Add files via upload
The "modern Olympics" comprises all the Games from Athens 1986 to Rio 2016. The Olympics is more than just a quadrennial multi-sport world championship. It is a lense through which to understand global history, including shifting geopolitical power dynamics, women’s empowerment, and the evolving values of society. In this kernel, my goal is to shed light on major patterns in Olympic history. How many athletes, sports, and nations are there? Where do most athletes come from? Who wins medals? What are the characteristic of the athletes (e.g., gender and physical size)?
These are the sort of tidbits I like to sprinkle in I scraped this data from www.sports-reference.com, which hosts a detailed database on Olympic history, which is developed and maintained by an independent group of Olympic.
Evaluate the evolution of the Olympic Games Exploratory data analysis can be used to evaluate the evolution of the Olympic Games over the years.
Improve performance Analysis can help member states improve their policies and strategies by providing them with the most recent statistics.
Find mistakesHost countries can use analysis to find mistakes in the arrangements of the event and overcome them.